+++ src/main/java/com/salesforce/storm/spout/dynamic/VirtualSpout.java
@@ -31,7 +31,7 @@ import com.google.common.collect.Maps;
 import com.salesforce.storm.spout.dynamic.consumer.Consumer;
 import com.salesforce.storm.spout.dynamic.consumer.ConsumerPeerContext;
 import com.salesforce.storm.spout.dynamic.consumer.Record;
-import com.salesforce.storm.spout.sideline.filter.FilterChain;
+import com.salesforce.storm.spout.dynamic.filter.FilterChain;
 import com.salesforce.storm.spout.dynamic.consumer.ConsumerState;
 import com.salesforce.storm.spout.dynamic.handler.VirtualSpoutHandler;
 import com.salesforce.storm.spout.dynamic.retry.RetryManager;
@@ -137,9 +137,8 @@ public class VirtualSpout implements DelegateSpout {
     private final Map<String, Long> ackTimeBuckets = Maps.newHashMap();
 
     /**
-     * Constructor.
-     * Use this constructor for your Sidelined instances.  IE an instance that has a specified starting and ending
-     * state.
+     * Create a new VirtualSpout instance.
+     *
      * @param virtualSpoutId Identifier for this VirtualSpout instance.
      * @param spoutConfig Our topology config
      * @param topologyContext Our topology context
@@ -191,7 +190,7 @@ public class VirtualSpout implements DelegateSpout {
             throw new IllegalStateException("Cannot call open more than once!");
         }
 
-        logger.info("Open has starting state {} and ending state {}", startingState, endingState);
+        logger.info("Open for VirtualSpout {} has starting state {} and ending state {}", getVirtualSpoutId(), startingState, endingState);
 
         // Create our failed msg retry manager & open
         retryManager = getFactoryManager().createNewFailedMsgRetryManagerInstance();
@@ -426,7 +425,7 @@ public class VirtualSpout implements DelegateSpout {
         }
         ackTimeBuckets.put("MessageId", ackTimeBuckets.get("MessageId") + (System.currentTimeMillis() - start));
 
-        // Talk to sidelineConsumer and mark the offset completed.
+        // Talk to Consumer and mark the offset completed.
         start = System.currentTimeMillis();
         consumer.commitOffset(messageId.getNamespace(), messageId.getPartition(), messageId.getOffset());
         ackTimeBuckets.put("CommitOffset", ackTimeBuckets.get("CommitOffset") + (System.currentTimeMillis() - start));