+++ src/main/java/com/salesforce/storm/spout/sideline/handler/SidelineSpoutHandler.java
@@ -33,15 +33,15 @@ import com.salesforce.storm.spout.dynamic.DynamicSpout;
 import com.salesforce.storm.spout.dynamic.FactoryManager;
 import com.salesforce.storm.spout.dynamic.handler.SpoutHandler;
 import com.salesforce.storm.spout.sideline.SidelineVirtualSpoutIdentifier;
-import com.salesforce.storm.spout.sideline.SpoutTriggerProxy;
 import com.salesforce.storm.spout.dynamic.VirtualSpout;
 import com.salesforce.storm.spout.dynamic.VirtualSpoutIdentifier;
 import com.salesforce.storm.spout.dynamic.kafka.KafkaConsumerConfig;
 import com.salesforce.storm.spout.dynamic.config.SpoutConfig;
 import com.salesforce.storm.spout.dynamic.consumer.ConsumerState;
 import com.salesforce.storm.spout.sideline.config.SidelineConfig;
-import com.salesforce.storm.spout.sideline.filter.FilterChainStep;
-import com.salesforce.storm.spout.sideline.filter.NegatingFilterChainStep;
+import com.salesforce.storm.spout.dynamic.filter.FilterChainStep;
+import com.salesforce.storm.spout.dynamic.filter.InvalidFilterChainStepException;
+import com.salesforce.storm.spout.dynamic.filter.NegatingFilterChainStep;
 import com.salesforce.storm.spout.sideline.persistence.PersistenceAdapter;
 import com.salesforce.storm.spout.sideline.persistence.SidelinePayload;
 import com.salesforce.storm.spout.sideline.trigger.SidelineRequest;
@@ -62,7 +62,7 @@ import java.util.Set;
 /**
  * Handler for managing sidelines on a DynamicSpout.
  */
-public class SidelineSpoutHandler implements SpoutHandler {
+public class SidelineSpoutHandler implements SpoutHandler, SidelineController {
 
     // Logger
     private static final Logger logger = LoggerFactory.getLogger(SidelineSpoutHandler.class);
@@ -71,6 +71,8 @@ public class SidelineSpoutHandler implements SpoutHandler {
      * Identifier for the firehose, or 'main' VirtualSpout instance.
      */
     private static final String MAIN_ID = "main";
+
+    private boolean isOpen = false;
 
     /**
      * The Spout configuration map.
@@ -102,6 +104,12 @@ public class SidelineSpoutHandler implements SpoutHandler {
      */
     @Override
     public void open(final Map<String, Object> spoutConfig) {
+        if (isOpen) {
+            throw new RuntimeException("SidelineSpoutHandler is already opened!");
+        }
+
+        isOpen = true;
+
         this.spoutConfig = spoutConfig;
 
         final String persistenceAdapterClass = (String) spoutConfig.get(SidelineConfig.PERSISTENCE_ADAPTER_CLASS);
@@ -115,6 +123,16 @@ public class SidelineSpoutHandler implements SpoutHandler {
             persistenceAdapterClass
         );
         persistenceAdapter.open(spoutConfig);
+    }
+
+    /**
+     * When this handler is closed this method is called.
+     */
+    @Override
+    public void close() {
+        persistenceAdapter.close();
+
+        isOpen = false;
     }
 
     /**
@@ -135,17 +153,34 @@ public class SidelineSpoutHandler implements SpoutHandler {
 
         createSidelineTriggers();
 
-        // Create the main spout for the namespace, we'll dub it the 'firehose'
-        fireHoseSpout = new VirtualSpout(
-            // We use a normal virtual spout identifier here rather than the sideline one because this is NOT a sideline, it's our firehose.
-            new DefaultVirtualSpoutIdentifier(getVirtualSpoutIdPrefix() + ":" + MAIN_ID),
-            getSpoutConfig(),
-            topologyContext,
-            spout.getFactoryManager(),
-            spout.getMetricsRecorder(),
-            null,
-            null
-        );
+        loadSidelines();
+
+        for (final SidelineTrigger sidelineTrigger : sidelineTriggers) {
+            sidelineTrigger.open(getSpoutConfig());
+        }
+    }
+
+    /**
+     * Loads existing sideline requests that have been previously persisted and checks to make sure that they are running.
+     */
+    private void loadSidelines() {
+        final VirtualSpoutIdentifier fireHoseIdentifier = new DefaultVirtualSpoutIdentifier(getVirtualSpoutIdPrefix() + ":" + MAIN_ID);
+
+        // If we haven't spun up a VirtualSpout yet, we create it here.
+        if (fireHoseSpout == null) {
+            // Create the main spout for the namespace, we'll dub it the 'firehose'
+            fireHoseSpout = new VirtualSpout(
+                // We use a normal virtual spout identifier here rather than the sideline one because this is NOT a sideline,
+                // it's our firehose.
+                fireHoseIdentifier,
+                getSpoutConfig(),
+                topologyContext,
+                spout.getFactoryManager(),
+                spout.getMetricsRecorder(),
+                null,
+                null
+            );
+        }
 
         final String topic = (String) getSpoutConfig().get(KafkaConsumerConfig.KAFKA_TOPIC);
 
@@ -161,11 +196,13 @@ public class SidelineSpoutHandler implements SpoutHandler {
             final Set<Integer> partitions = getPersistenceAdapter().listSidelineRequestPartitions(id);
 
             for (final Integer partition : partitions) {
-                payload = getPersistenceAdapter().retrieveSidelineRequest(id, partition);
+                payload = retrieveSidelinePayload(id, partition);
 
                 if (payload == null) {
+                    logger.warn("Sideline {} on partition {} payload was null, this is probably a serialization problem.");
                     continue;
                 }
+
                 startingStateBuilder.withPartition(topic, partition, payload.startingOffset);
 
                 // We only have an ending offset on STOP requests
@@ -175,7 +212,7 @@ public class SidelineSpoutHandler implements SpoutHandler {
             }
 
             if (payload == null) {
-                logger.warn("Sideline request {} did not have any partitions persisted", id);
+                logger.warn("Sideline request {} did not have any persisted partitions", id);
                 continue;
             }
 
@@ -183,15 +220,21 @@ public class SidelineSpoutHandler implements SpoutHandler {
             if (payload.type.equals(SidelineType.START)) {
                 logger.info("Resuming START sideline {} {}", payload.id, payload.request.step);
 
-                fireHoseSpout.getFilterChain().addStep(
-                    payload.id,
-                    payload.request.step
-                );
+                // Only add the step if the id isn't already in the chain.  Note that we do NOT check for the step here, it's possible
+                // though very unusual to have the exact same step used in different requests. While that's silly, we don't want to choke
+                // on it here and put the filter chain in a weird state.
+                if (!fireHoseSpout.getFilterChain().hasStep(payload.id)) {
+                    fireHoseSpout.getFilterChain().addStep(
+                        payload.id,
+                        payload.request.step
+                    );
+                }
             }
 
             // Resuming a stopped request means we spin up a new sideline spout
             if (payload.type.equals(SidelineType.STOP)) {
-                openVirtualSpout(
+                // This method will check to see that the VirtualSpout isn't already in the SpoutCoordinator before adding it
+                addSidelineVirtualSpout(
                     payload.id,
                     payload.request.step,
                     startingStateBuilder.build(),
@@ -200,14 +243,12 @@ public class SidelineSpoutHandler implements SpoutHandler {
             }
         }
 
-        for (final SidelineTrigger sidelineTrigger : sidelineTriggers) {
-            sidelineTrigger.open(getSpoutConfig());
-        }
-
         // After altering the filter chain is complete, lets NOW start the fire hose
         // This keeps a race condition where the fire hose could start consuming before filter chain
         // steps get added.
-        spout.addVirtualSpout(fireHoseSpout);
+        if (!spout.hasVirtualSpout(fireHoseIdentifier)) {
+            spout.addVirtualSpout(fireHoseSpout);
+        }
     }
 
     /**
@@ -228,11 +269,19 @@ public class SidelineSpoutHandler implements SpoutHandler {
     }
 
     /**
+     * Does a sideline exist in the started state?
+     * @param sidelineRequest sideline request.
+     * @return true it does, false it does not.
+     */
+    public boolean isSidelineStarted(SidelineRequest sidelineRequest) {
+        return fireHoseSpout.getFilterChain().hasStep(sidelineRequest.id);
+    }
+
+    /**
      * Starts a sideline request.
      * @param sidelineRequest representation of the request that is being started.
-     * @return sideline request identifier.
      */
-    public SidelineRequestIdentifier startSidelining(SidelineRequest sidelineRequest) {
+    public void startSidelining(SidelineRequest sidelineRequest) {
         logger.info("Received START sideline request");
 
         // Store the offset that this request was made at, when the sideline stops we will begin processing at
@@ -256,8 +305,17 @@ public class SidelineSpoutHandler implements SpoutHandler {
 
         // Update start count metric
         spout.getMetricsRecorder().count(getClass(), "start-sideline", 1L);
+    }
 
-        return sidelineRequest.id;
+    /**
+     * Does a sideline exist in the stopped state?
+     * @param sidelineRequest sideline request.
+     * @return true it has, false it has not.
+     */
+    public boolean isSidelineStopped(SidelineRequest sidelineRequest) {
+        return spout.hasVirtualSpout(
+            generateSidelineVirtualSpoutId(sidelineRequest.id)
+        );
     }
 
     /**
@@ -265,7 +323,7 @@ public class SidelineSpoutHandler implements SpoutHandler {
      * @param sidelineRequest A representation of the request that is being stopped
      */
     public void stopSidelining(SidelineRequest sidelineRequest) {
-        final SidelineRequestIdentifier id = fireHoseSpout.getFilterChain().findStep(sidelineRequest.step);
+        final SidelineRequestIdentifier id = (SidelineRequestIdentifier) fireHoseSpout.getFilterChain().findStep(sidelineRequest.step);
 
         if (id == null) {
             logger.error(
@@ -280,7 +338,7 @@ public class SidelineSpoutHandler implements SpoutHandler {
         logger.info("Received STOP sideline request");
 
         // Remove the steps associated with this sideline request
-        final FilterChainStep step = fireHoseSpout.getFilterChain().removeSteps(id);
+        final FilterChainStep step = fireHoseSpout.getFilterChain().removeStep(id);
         // Create a negated version of the step we just pulled from the firehose
         final FilterChainStep negatedStep = new NegatingFilterChainStep(step);
 
@@ -294,10 +352,15 @@ public class SidelineSpoutHandler implements SpoutHandler {
         // assigned to this particular sideline spout instance
         for (final ConsumerPartition consumerPartition : endingState.getConsumerPartitions()) {
             // This is the state that the VirtualSpout should start with
-            final SidelinePayload sidelinePayload = getPersistenceAdapter().retrieveSidelineRequest(
+            final SidelinePayload sidelinePayload = retrieveSidelinePayload(
                 id,
                 consumerPartition.partition()
             );
+
+            if (sidelinePayload == null) {
+                logger.warn("Sideline {} on partition {} payload was null, this is probably a serialization problem.");
+                continue;
+            }
 
             logger.info("Loaded sideline payload for {} = {}", consumerPartition, sidelinePayload);
 
@@ -318,7 +381,7 @@ public class SidelineSpoutHandler implements SpoutHandler {
         // Build our starting state, this is a map of partition and offset
         final ConsumerState startingState = startingStateBuilder.build();
 
-        openVirtualSpout(
+        addSidelineVirtualSpout(
             id,
             negatedStep,
             startingState,
@@ -376,7 +439,7 @@ public class SidelineSpoutHandler implements SpoutHandler {
      * @param startingState Starting consumer state
      * @param endingState Ending consumer state
      */
-    private void openVirtualSpout(
+    private void addSidelineVirtualSpout(
         final SidelineRequestIdentifier id,
         final FilterChainStep step,
         final ConsumerState startingState,
@@ -384,6 +447,11 @@ public class SidelineSpoutHandler implements SpoutHandler {
     ) {
         // Generate our virtualSpoutId using the payload id.
         final VirtualSpoutIdentifier virtualSpoutId = generateSidelineVirtualSpoutId(id);
+
+        if (spout.hasVirtualSpout(virtualSpoutId)) {
+            logger.info("VirtualSpout {} is already running", virtualSpoutId);
+            return;
+        }
 
         // This info is repeated in VirtualSpout.open(), not needed here.
         logger.debug("Starting VirtualSpout {} with starting state {} and ending state", virtualSpoutId, startingState, endingState);
@@ -430,9 +498,18 @@ public class SidelineSpoutHandler implements SpoutHandler {
         return new SidelineVirtualSpoutIdentifier(getVirtualSpoutIdPrefix(), sidelineRequestIdentifier);
     }
 
+    private SidelinePayload retrieveSidelinePayload(final SidelineRequestIdentifier id, final int partition) {
+        try {
+            return getPersistenceAdapter().retrieveSidelineRequest(id, partition);
+        } catch (InvalidFilterChainStepException ex) {
+            logger.error("Unable to load sideline payload {}", ex);
+            // Basically if we can't deserialize the step we're not sending back any part of the payload.
+            return null;
+        }
+    }
+
     /**
-     * Create an instance of the configured StartingTrigger.
-     * @return Instance of a StartingTrigger
+     * Create an instance of the configured SidelineTrigger.
      */
     @SuppressWarnings("unchecked")
     synchronized void createSidelineTriggers() {
@@ -451,8 +528,8 @@ public class SidelineSpoutHandler implements SpoutHandler {
             // Will throw a RuntimeException if this is not configured correctly
             final SidelineTrigger sidelineTrigger = FactoryManager.createNewInstance(sidelineTriggerClass);
 
-            // Revisit the purpose of this proxy
-            sidelineTrigger.setSidelineSpout(new SpoutTriggerProxy(this));
+            // This is the new preferred approach for triggers
+            sidelineTrigger.setSidelineController(this);
 
             // Add it to our collection so we can iterate it later.
             sidelineTriggers.add(sidelineTrigger);
@@ -485,9 +562,5 @@ public class SidelineSpoutHandler implements SpoutHandler {
 
     PersistenceAdapter getPersistenceAdapter() {
         return persistenceAdapter;
-    }
-
-    private FactoryManager getFactoryManager() {
-        return new FactoryManager(spoutConfig);
     }
 }